requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
astor=0.8.1=pypi_0
ca-certificates=2022.9.14=ha878542_0
certifi=2022.9.14=pyhd8ed1ab_0
charset-normalizer=2.1.1=pypi_0
docopt=0.6.2=pypi_0
gast=0.5.3=pypi_0
google-pasta=0.2.0=pypi_0
googledrivedownloader=0.4=pypi_0
grpcio=1.48.1=pypi_0
h5py=3.7.0=pypi_0
hhsuite=3.3.0=py37pl5321h675a0cb_5
idna=3.4=pypi_0
importlib-metadata=4.12.0=pypi_0
isodate=0.6.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnsl=2.0.0=h7f98852_0
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
ncurses=6.3=h5eee18b_3
networkx=2.6.3=pypi_0
numpy=1.21.6=pypi_0
openssl=1.1.1q=h7f8727e_0
pandas=1.3.5=pypi_0
perl=5.32.1=2_h7f98852_perl5
pillow=9.2.0=pypi_0
pip=22.1.2=py37h06a4308_0
pipreqs=0.4.11=pypi_0
protobuf=3.20.1=pypi_0
pyparsing=3.0.9=pypi_0
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
python-louvain=0.16=pypi_0
python_abi=3.7=2_cp37m
pytz=2022.2.1=pypi_0
pyyaml=6.0=pypi_0
rdflib=6.2.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
setuptools=63.4.1=py37h06a4308_0
six=1.16.0=pypi_0
sqlite=3.39.2=h5082296_0
tensorboard=1.14.0=pypi_0
tensorboardx=2.5.1=pypi_0
tensorflow=1.14.0=pypi_0
tensorflow-estimator=1.14.0=pypi_0
termcolor=2.0.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.11.0+cu113=pypi_0
torch-cluster=1.6.0=pypi_0
torch-geometric=1.7.2=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.15=pypi_0
torch-spline-conv=1.2.1=pypi_0
torchaudio=0.11.0+cu113=pypi_0
torchvision=0.12.0+cu113=pypi_0
tqdm=4.64.1=pypi_0
typing-extensions=4.3.0=pypi_0
urllib3=1.26.12=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.5=h7f8727e_1
yarg=0.1.9=pypi_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h5eee18b_3
